+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Option to add unit labels by passing a mapping in QuantityField unit_options
+
 ## [0.7.0] - 2024-10-21
 ### Added
 - Quantity field for (value, unit) pairs, allowing auto-conversion for standard units